Add files via upload
Build instruction- we have used to shuffle min and max numbers the command shuff is used
usage- ./random.sh
description- make a function called function roll_num and here the min and max no. are 1 and 10 in script the n is integer and the output store in the variable to print that variable we use echo $number
limitation- these are the hard-code values and if put float or any minus no. we get erros.